Work queue model for loaders
Summary:
This diff addresses the tail off in loading throughput towards the
end of the loading process due to skew in the chunk sizes given to the
loader threads. Because of the distribution of links, some chunks of the
id1 space have many more links to load than others.  The most robust
approach I could think of to solve this was to have a smaller pool of
loader threads (~10) and dynamically assign the chunks to threads with
a work queue.  If we break the work up into chunks such that
least work, then all of the threads will likely run out of work at
nearly the same time.

Since the per-thread progress reports aren't meaningful in this model,
I added in a global progress report and have chunk progress reports if
debug logging is turned on.

Task ID: #1143673

Blame Rev:

Test Plan: Ran loader for 10e7 id1s and checked table sizes and top counts for sanity

Reviewers: vamsi, dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D3921