Add in historical range queries.
Summary:
Before this diff, only the first 10000 assocs for big assoc were every
retrieved.

This change simulates the process of the client fetching the first 10000
assocs, then selecting the next 10000 based on timestamp.

A small list of bounded size is maintained in the Requester of (id1/assoc_type) pairs
where the 10000 row limit was hit, along with the timestamp of the last
assoc.  This data can then be used to construct queries for the further
back history.

The % of get_linklist requests which request history data (not the most
recent rows) is controlled by a config key "getlinklist_history"

Added the ability to change the row limit from the default (10000) to
another value.  This is used for unit testing currently, but in future
could be used to experiment with different row limits.

Test Plan:
ant test

New unit test added that exercises requester history query code path.

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D4941