Variable data size for nodes and links
Summary:
Previously all links/nodes had same data length.  This isn't very
realistic, so this diff makes the data length variable.  The
distribution is log-normal, which has a similar shape to the observed
real distribution.  The median size values in the config file give a
distribution with the same mean size as the real distribution.

Add log-normal distribution.

Refactor the initialisation for the request somewhat since it was
getting really unwieldy with additions.

Test Plan: ant test <-- all tests pass

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D5121