Fix id2 selection so that we get expected hits/misses
Summary:
We want the operation id2s to be chosen so that link updates and deletes are mostly for existing (id1, id2) links, and so that link inserts are mostly for non-existent (id1, id2) links.  Add test and fix bug so that this is the case.

Refactor so that nlinks calculation in one place rather than two

Stop startid1 being silently changed

Test Plan:
ant test passes

Added test to make sure behaves as expecte

ran test, checked that expected number of

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D5157