Multiple link types for LinkBench
Summary:
Add in config parameter to control number of link types (default to 1 if
not provided).

The number of links per id1 is the same as before the change, but these
links are divided between the n link types.  If we number the link types
1 to n, the links are distributed in a round robin fashion.  E.g. if
there are five types and two links for id1=1234, then there is a link of type 1 and a
link of type 2.  If there are five types, there is a link of each type.

This change makes the count table much bigger.  Switch bulk count add
logic to use "REPLACE INTO" to speed up loading of link counts.

Test Plan:
ant test

Did a small benchmark run to check that average range scan size was
reduced proportionally (it was)

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D5253