New Motif data generator
Summary:
Try to generate somewhat more realistic data that has some of the
compressibility properties of real data.  The basic idea is that
a data generator has a set of "motifs" that will repeatedly occur
in the output data.  This is implemented by having an internal buffer of
random data from which fragments are inserted into the output.

Test Plan:
Unit test to exercise code.
Test the compressibility of the output by compressing with gzip to make
sure that the compreesibility of the output data is being changed in the
way we'd expect as parameters vary.

Test the performance of the data generator to make sure it won't be
a performance bottleneck.  It seems like it is at worse marginally slower
than the uniform data generator, and at best somewhat faster (if there
are many repeated motifs).  In either case it won't be a bottleneck.

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D5289