Fix LinkBench to not depend on Mysql bug 46675
Summary:
Previously LinkBench depended on the behaviour identified in Mysql bug 46675
for determining how to update link counts.  Recent versions of MySql (both on the 5.5 and 5.6 releases) have had this bug fixed.

This revision fixes the problem by using the affected rows count, which can distinguish between the three cases.

Test Plan:
ant test passes on mysql server versions 5.5.24-ubuntu and 5.1.53-faceboo
(with and without the bugfix in question)

Reviewers: dhruba, vamsi

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D6009