Remove line causing the database to close twice and fault
Summary:
Removed linkStore.close line that was added before closeStores was introduced.
This is what was causing linkbench for rocksdb to not run successfully with multiple
threads. Now it runs for both load and request phases with multiple threads.

Test Plan:
ant dist; load phase with loaders=10, request phase with requesters=100
and number of requests=500000

Reviewers: dhruba, tarmstrong, vamsi, sheki

Reviewed By: sheki

Differential Revision: https://reviews.facebook.net/D8793